pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-on-mobile"
version = "0.1.0"
description = "Cross-platform mobile-friendly CLI to list, download, and run open-source LLMs (Android Termux + iOS via Pythonista/Shortcuts)."
requires-python = ">=3.9"
readme = "README.md"
authors = [{ name = "Your Name" }]
license = { text = "MIT" }
keywords = ["llm", "mobile", "android", "ios", "termux", "offline"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "requests>=2.31.0",
  "tqdm>=4.66.0",
  "typer>=0.12.0",
  "rich>=13.7.0",
  "huggingface_hub>=0.21.4",
]

[project.optional-dependencies]
llama = ["llama-cpp-python>=0.2.90"]

[project.scripts]
lom = "llm_on_mobile.cli:app"
